add data from the pdf

+++ src/main/java/tr/com/oderopay/model/garantipay/request/GarantiPayInitRequestDataPrice.java
@@ -9,5 +9,5 @@ import lombok.Data;
 @Builder
 public class GarantiPayInitRequestDataPrice {
     private String amount;
-    private Double currency;
+    private Integer currency;
 }
